builder_image/requirements.txt
ipdb
ansible
python-novaclient
python-glanceclient
python-keystoneclient
# neutronclient
